requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.4
feedparser==6.0.11
fastapi==0.115.12
jinja2==3.1.6
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
uvicorn==0.34.0
